refactor: Remove unused backup components
