--- frontend/src/app/projects/project-detail/backup-settings/backup-settings.component.css
@@ -1,8 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-.content {
-  margin: 10px;
-}
