--- frontend/src/app/projects/project-detail/backup-settings/backup-settings.component.html
@@ -1,21 +1,0 @@
-<!--
- ~ SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- ~ SPDX-License-Identifier: Apache-2.0
- -->
-
-<mat-card
-  ><h2>Backups</h2>
-  <mat-tab-group animationDuration="0" mat-align-tabs="center">
-    <mat-tab label="EASE"
-      ><div class="content">
-        <app-ease-backup-settings
-          [project]="project"
-        ></app-ease-backup-settings>
-      </div>
-    </mat-tab>
-    <mat-tab label="Jenkins">
-      <div class="content">
-        <app-jenkins-backup-settings></app-jenkins-backup-settings>
-      </div>
-    </mat-tab> </mat-tab-group
-></mat-card>
