--- frontend/src/app/projects/project-detail/backup-settings/backup-settings.component.spec.ts
@@ -1,29 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { BackupSettingsComponent } from './backup-settings.component';
-
-describe('BackupSettingsComponent', () => {
-  let component: BackupSettingsComponent;
-  let fixture: ComponentFixture<BackupSettingsComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [BackupSettingsComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BackupSettingsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
