--- frontend/src/app/projects/project-detail/backup-settings/backup-settings.component.ts
@@ -1,20 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { Component, Input, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-backup-settings',
-  templateUrl: './backup-settings.component.html',
-  styleUrls: ['./backup-settings.component.css'],
-})
-export class BackupSettingsComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-
-  @Input()
-  project: string = '';
-}
