--- frontend/src/app/projects/project-detail/backup-settings/jenkins-backup-settings/jenkins-backup-settings.component.css
@@ -1,72 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-button {
-  margin: 5px 10px 0 0;
-}
-
-input:disabled {
-  color: black;
-}
-
-.state {
-  padding: 0px;
-  width: 201px;
-  border-radius: 3px 3px 0 0;
-}
-
-.success {
-  background-color: green;
-}
-
-.failure {
-  background-color: darkred;
-}
-
-.unknown {
-  background-color: darkblue;
-}
-
-.inner-state {
-  padding: 10.5px 10.5px 0 10.5px;
-}
-
-.inner-state > .text {
-  font-size: 14px;
-  padding-top: 3.5px;
-  padding-bottom: 10.5px;
-  line-height: 15px;
-  display: inline-block;
-  color: white;
-}
-
-.inner-state > .title {
-  color: lightgrey;
-  font-size: 11px;
-  line-height: 11.813px;
-}
-
-#state-field {
-  justify-content: space-between;
-  align-items: center;
-}
-
-#error-icon {
-  margin-right: 10px;
-  color: white;
-}
-
-.spinner {
-  float: right;
-}
-
-#refresh-button {
-  float: right;
-  margin: -8px -8px;
-}
-
-h3 {
-  display: inline;
-}
