--- frontend/src/app/projects/project-detail/backup-settings/jenkins-backup-settings/jenkins-backup-settings.component.html
@@ -1,72 +1,0 @@
-<!--
- ~ SPDX-FileCopyrightText: Copyright DB Netz AG and the capella-collab-manager contributors
- ~ SPDX-License-Identifier: Apache-2.0
- -->
-
-This option is currently not available.
-<!--
-    <div class="jenkins">
-  <h3>Available Backup Jobs</h3>
-  <mat-spinner *ngIf="loading" class="spinner" [diameter]="25"></mat-spinner>
-  <button
-    id="refresh-button"
-    (click)="refreshPipeline()"
-    mat-icon-button
-    *ngIf="!loading && pipeline"
-  >
-    <app-mat-icon>refresh</app-mat-icon>
-  </button>
-  <div>
-    <div *ngIf="!loading">No backup jobs were found. <br /></div>
-    <button mat-flat-button color="primary" (click)="createPipeline()">
-      Create Backup Job
-    </button>
-  </div>
-  <div *ngIf="pipeline">
-    <br />
-    <div class="">
-      <div class="flexbox">
-        <mat-form-field appearance="fill">
-          <mat-label>Name of the Backup Job</mat-label>
-          <input matInput disabled [value]="pipeline.name" /> </mat-form-field
-        ><br />
-        <div
-          class="state flexbox"
-          [ngClass]="pipeline.latest_run.result.toLowerCase()"
-          id="state-field"
-          *ngIf="pipeline.latest_run"
-        >
-          <div class="inner-state">
-            <div class="title">
-              Last Run: {{ pipeline.latest_run.start_time }}
-            </div>
-            <span class="text">{{ pipeline.latest_run.result }}</span
-            ><br />
-          </div>
-          <app-mat-icon
-            *ngIf="pipeline.latest_run.result == 'FAILURE'"
-            id="error-icon"
-            >error</app-mat-icon
-          >
-        </div>
-      </div>
-    </div>
-
-    <button mat-flat-button (click)="runJenkinsJob()" color="primary">
-      Run job
-    </button>
-    <button mat-flat-button (click)="removeJenkinsPipeline()" color="primary">
-      Remove Backup Pipeline
-    </button>
-    <a
-      mat-flat-button
-      color="primary"
-      target="_blank"
-      [href]="pipeline.latest_run.logs_url"
-      *ngIf="pipeline.latest_run"
-    >
-      View logs <app-mat-icon>launch</app-mat-icon>
-    </a>
-  </div>
-</div>
--->
