+++ frontend/src/app/projects/project-detail/model-overview/model-overview.component.html
@@ -52,7 +52,7 @@
             <app-mat-icon>sync</app-mat-icon>
           </a>
           <a
-            *ngIf="integrations.git && model.git_model"
+            *ngIf="model.git_model"
             mat-mini-fab
             [disabled]="!model.git_model"
             matTooltip="Open git repository"
@@ -60,7 +60,7 @@
             <app-mat-icon>link</app-mat-icon>
           </a>
           <a
-            *ngIf="integrations.t4c && model.t4c_model"
+            *ngIf="model.t4c_model"
             routerLink="/workspaces"
             mat-mini-fab
             matTooltip="Request persistent session"
@@ -69,7 +69,7 @@
           </a>
           <a
             mat-mini-fab
-            *ngIf="integrations.git && model.git_model"
+            *ngIf="model.git_model"
             matTooltip="Open readonly session"
           >
             <app-mat-icon>remove_red_eye</app-mat-icon>
